+++ checker/src/main/java/org/checkerframework/checker/formatter/FormatterAnnotatedTypeFactory.java
@@ -86,7 +86,7 @@ public class FormatterAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
    * from its first argument.
    */
   @Override
-  public void prepareMethodForWriting(AMethod method) {
+  public void wpiPrepareMethodForWriting(AMethod method) {
     if (hasFormatMethodAnno(method)) {
       AField param = method.parameters.get(0);
       if (param != null) {
@@ -104,7 +104,7 @@ public class FormatterAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
    * from its first argument.
    */
   @Override
-  public void prepareMethodForWriting(
+  public void wpiPrepareMethodForWriting(
       WholeProgramInferenceJavaParserStorage.CallableDeclarationAnnos methodAnnos) {
     if (hasFormatMethodAnno(methodAnnos)) {
       AnnotatedTypeMirror atm = methodAnnos.getParameterType(0);